requirements.txt
databricks-connect==16.0.0
databricks-sdk[openai]==0.46.0
databricks-sql-connector==4.0.0
pandas==2.2.3
streamlit==1.41.1
python-dotenv
plotly
sqlalchemy
psycopg[binary]; platform_machine != "arm64"
psycopg; platform_machine == "arm64"  # For Apple Silicon, install libpq/OpenSSL locally and build from source
pytest